API 2.9.4: Fix Bundle Fullness predicate, expose playFabId in GeyserConnection (#6190)
* API 2.9.4

* Deprecate bundleFullness(int), introduce bundleFullness(double)

* Allow overriding vanilla item model without additional predicates

* Expose playFabId in GeyserConnection

* Address clanker review

+++ core/src/main/java/org/geysermc/geyser/registry/loader/ProviderRegistryLoader.java
@@ -213,7 +213,7 @@ public class ProviderRegistryLoader implements RegistryLoader<Map<Class<?>, Prov
         int length = args.length;
         switch (property) {
             case BUNDLE_FULLNESS -> {
-                return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.BUNDLE_FULLNESS, (int) args[1]);
+                return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.BUNDLE_FULLNESS, (double) args[1]);
             }
             case DAMAGE -> {
                 // One with, one without normalization
